// packages/wcag-data/src/types.ts
export interface WcagDataDirectories {
   packageRoot: string;
   raw: string;
   generated: string;
   scripts: string;
   test: string;
}